CODA-269 - Refactor package naming
